Move optimized* classes into compiler namespace
