+++ src/Compiler/OptimizedRewriteRule.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace ToyWpRouting;
+namespace ToyWpRouting\Compiler;
+
+use ToyWpRouting\RewriteRuleInterface;
 
 class OptimizedRewriteRule implements RewriteRuleInterface
 {
