+++ tests/fixtures/rewrite-cache.php
@@ -11,11 +11,11 @@ return function (?\ToyWpRouting\InvocationStrategyInterface $invocationStrategy 
         {
             parent::__construct('', $invocationStrategy);
 
-            $this->rewrites->attach(new \ToyWpRouting\OptimizedRewrite(array (
+            $this->rewrites->attach(new \ToyWpRouting\Compiler\OptimizedRewrite(array (
   0 => 'GET',
   1 => 'HEAD',
 ), array (
-  0 => new \ToyWpRouting\OptimizedRewriteRule('9f79cebcf1735d5eaefeee8dbc7316dd', 'index.php?var=first&matchedRule=9f79cebcf1735d5eaefeee8dbc7316dd', array (
+  0 => new \ToyWpRouting\Compiler\OptimizedRewriteRule('9f79cebcf1735d5eaefeee8dbc7316dd', 'index.php?var=first&matchedRule=9f79cebcf1735d5eaefeee8dbc7316dd', array (
   'var' => 'var',
   'matchedRule' => 'matchedRule',
 ), '^first$'),
@@ -23,10 +23,10 @@ return function (?\ToyWpRouting\InvocationStrategyInterface $invocationStrategy 
   'var' => 'var',
   'matchedRule' => 'matchedRule',
 ), $this->invocationStrategy, 'firsthandler', NULL));
-$this->rewrites->attach(new \ToyWpRouting\OptimizedRewrite(array (
+$this->rewrites->attach(new \ToyWpRouting\Compiler\OptimizedRewrite(array (
   0 => 'POST',
 ), array (
-  0 => new \ToyWpRouting\OptimizedRewriteRule('3cf5d427e03a68a3881d2d68a86b64f1', 'index.php?var=second&matchedRule=3cf5d427e03a68a3881d2d68a86b64f1', array (
+  0 => new \ToyWpRouting\Compiler\OptimizedRewriteRule('3cf5d427e03a68a3881d2d68a86b64f1', 'index.php?var=second&matchedRule=3cf5d427e03a68a3881d2d68a86b64f1', array (
   'var' => 'var',
   'matchedRule' => 'matchedRule',
 ), '^second$'),
